requirements.txt
# Flask Framework
Flask==2.3.3
flask-cors==4.0.0

# OpenCV untuk image processing (headless untuk server)
opencv-python-headless==4.8.0.76

# YOLOv5 Dependencies (optimized untuk production)
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
Pillow==10.0.0

# Untuk YOLOv5 inference
ultralytics==8.0.147

# Cloudinary SDK
cloudinary==1.33.0

# Firebase Admin SDK
firebase-admin==6.2.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Image Processing & Utils
scikit-image==0.21.0
matplotlib==3.7.2

# API Testing
werkzeug==2.3.7

# Production server untuk Railway
gunicorn==21.2.0